Consolidate USB utils into the USB integration
