+++ homeassistant/components/homeassistant_sky_connect/config_flow.py
@@ -16,6 +16,8 @@ from homeassistant.components.homeassistant_hardware.helpers import (
 from homeassistant.components.homeassistant_hardware.util import (
     ApplicationType,
     FirmwareInfo,
+)
+from homeassistant.components.usb import (
     usb_service_info_from_device,
     usb_unique_id_from_service_info,
 )
@@ -173,6 +175,7 @@ class HomeAssistantSkyConnectConfigFlow(
         self, fw_discovery_info: HardwareFirmwareDiscoveryInfo
     ) -> ConfigFlowResult:
         """Handle import from ZHA/OTBR firmware notification."""
+        assert fw_discovery_info["usb_device"] is not None
         usb_info = usb_service_info_from_device(fw_discovery_info["usb_device"])
         unique_id = usb_unique_id_from_service_info(usb_info)
